CBG-478 Add basic JSONMarshal/JSONUnmarshal wrappers
CBG-478 Add NewDecoder/NewEncoder wrappers and common interface
