+++ base/util_ce.go
@@ -1,0 +1,42 @@
+// +build !cb_sg_enterprise
+
+package base
+
+import (
+	"encoding/json"
+	"fmt"
+	"io"
+)
+
+// ErrDeltasNotSupported is returned when these functions are called in CE
+var ErrDeltasNotSupported = fmt.Errorf("Deltas not supported in CE")
+
+// Diff is only implemented in EE, the CE stub always returns an error.
+func Diff(old, new map[string]interface{}) (delta []byte, err error) {
+	return nil, ErrDeltasNotSupported
+}
+
+// Patch is only implemented in EE, the CE stub always returns an error.
+func Patch(old *map[string]interface{}, delta map[string]interface{}) (err error) {
+	return ErrDeltasNotSupported
+}
+
+// JSONUnmarshal parses the JSON-encoded data and stores the result in the value pointed to by v.
+func JSONUnmarshal(data []byte, v interface{}) error {
+	return json.Unmarshal(data, v)
+}
+
+// JSONMarshal returns the JSON encoding of v.
+func JSONMarshal(v interface{}) ([]byte, error) {
+	return json.Marshal(v)
+}
+
+// JSONDecoder returns a new JSON decoder implementing the JSONDecoderI interface
+func JSONDecoder(r io.Reader) JSONDecoderI {
+	return json.NewDecoder(r)
+}
+
+// JSONEncoder returns a new JSON encoder implementing the JSONEncoderI interface
+func JSONEncoder(w io.Writer) JSONEncoderI {
+	return json.NewEncoder(w)
+}
